+++ jaorm-dsl/src/main/java/io/github/ulisse1996/jaorm/dsl/query/common/MergeEnd.java
@@ -1,0 +1,6 @@
+package io.github.ulisse1996.jaorm.dsl.query.common;
+
+public interface MergeEnd<T> {
+
+    void execute();
+}