+++ jaorm-dsl/src/main/java/io/github/ulisse1996/jaorm/dsl/query/common/MergeEndNotMatching.java
@@ -1,0 +1,6 @@
+package io.github.ulisse1996.jaorm.dsl.query.common;
+
+import io.github.ulisse1996.jaorm.dsl.query.common.trait.WithMatching;
+
+public interface MergeEndNotMatching<T> extends WithMatching<T>, MergeEnd<T> {
+}