+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 1.7.0
+- Feature: SL4J support for standard Logging with custom fallback on SPI JaormLoggerHandler
+- Feature: List\<TableRow> support for @Query
+- Feature: Merge DSL
+
 ## 1.6.0
 - Refactor: Move generation bundle SPI to SPI Provider for Single Feature
 - Bugfix: Fix quarkus dev reload for class checks